requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.6.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.25.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
